Upgrading Webpack v4 -> v5
Also upgrading relevant Webpack plugin packages to the latest
versions.

`optimize-css-assets-webpack-plugin` is being removed because it
apparently doesn't support Webpack 5. Its docs say to use
`css-minimizer-webpack-plugin` instead.

The `TerserPlugin` configuration is being changed so that the
existing Webpack 4 behavior of not minifying JavaScript files in
the `lib/` folder and not extracting license comments out to a
separate `jssynth.js.LICENSE.txt` file is preserved.