Simpler logic for detecting touched piano key
Each piano key contains an inner element that displays the name of
the key (such as "B3" or "D#5/Eb5") and possibly also a "Root"
label if the instrument is a sampler instrument. When detecting
which piano key is under a touch or the mouse cursor, this inner
element needs to be accounted for, so that it is not chosen in
favor of the parent piano key element.

Previously, this was handled by special JavaScript logic. However,
if pointer events are disabled on these elements via CSS, then
they are "invisible" to `document.elementFromPoint()`, and the
special JavaScript logic is not needed. This seems a little better
because it removes the need for the JavaScript code to rely on a
particular DOM structure.