Simpler way of aligning piano keys
Adding an appropriate negative right margin to black piano keys
has the same effect as an equivalent negative left margin on the
relevant white piano keys, but has the advantage of not requiring
an extra CSS class to be present on the white piano keys.
(Although in practice all piano keys had a class indicating the
note name for this purpose, only the classes for the white A, B,
D, E, and G keys actually had any effect).