More locally defined CSS style
Since the purpose of this style is to horizontally center keyboard
key note labels, it seems to be more clear to place this style
directly on the relevant note label CSS class, rather than on an
ancestor container element.

While making this change I realized that two different styles are
needed to center the piano key labels, so I'm also adding comments
to make it easier to know that in the future.