Removing unneeded wrapper `<div>`
By combining the `.keyboard-container` and `.keyboard` CSS classes
there is only a need for a single wrapping `<div>` around each set
of piano key elements, not two. As a side benefit, this removes
the need to set an explicit element height (which was only needed
for the removed inner `<div className="keyboard">` element).