Laying out piano keys with Flexbox
This implementation seems easier to understand than the previous
version, in part because it allows removing several styles.

For future reference, the removed `white-space: nowrap` style was
needed to allow overflow from the keyboard keys container to be
horizontally scrollable (instead of wrapping). The `overflow-y:
hidden` style was needed to hide vertical overflow at the bottom
of the keyboard key container which isn't present in the new
implementation.