Removing unneeded `position: relative` style
There shouldn't be any absolutely positioned elements that rely on
an element with this class as a parent/ancestor, and removing this
style doesn't seem to cause any visual changes. Although the
absolutely positioned piano key labels _do_ require a relatively
positioned parent, that role is filled by elements with the
`.keyboard-key` class.