Handling `touchcancel` events in piano keyboards
Previously, when a `touchcancel` event was triggered for an active
touch on a piano keyboard key, the key would continue to be
highlighted and the note would play forever, with no way to stop
it except for navigating to a different view. If the touch was on
the left or right keyboard scroll button, the button would become
"stuck" and the keyboard would scroll in that direction forever.
If you then navigated to a different view and returned back, an
error would be repeatedly logged in the console forever.

Conceptually it seems like a `touchcancel` event should be treated
the same as a `touchend` event. That is, the same as if the touch
was moved or lifted off the element. Therefore, this commit does
just that by using the same event handler function for both
events.