Upgrading React v17.0.2 -> v18.2.0
The `flushSync()` calls are needed to work around React 18's new
automatic batching of state updates. Without them, some logic that
uses the current component state will not see the correct up to
date state.

The biggest impact of the `flushSync()` calls not being present is
that upon app initialization the default song won't be properly
synced to the synth layer, leading to silence during song playback
and errors when interacting with the app. Piano keyboards will
also be affected, going into a bad state when dragging from a
piano key onto one of the scroll buttons.