Fixing erroneous piano key re-triggering
The React upgrade in 59da53e849c547c7353232df919a986ef28b82b6
(v17 -> v18) introduced a bug that wasn't caught in that commit.
After triggering a piano keyboard key with the mouse, dragging off
the key with the mouse button down, releasing the mouse button,
and then moving the mouse cursor back onto the keyboard with the
mouse button released, the piano key (or scroll button) under the
cursor would be erroneously re-triggered.

The problem was a missing `flushSync()` call, needed when using
React v18. Without it, a state update in a `mouseover` event to
set the piano keyboard to "deactivated" state will not be visible
to an immediately subsequent `mousemove` event, causing the issue.