Preventing "pull down to refresh" in iOS/iPadOS
This disables the "pull down to refresh" functionality first added
in iOS/iPadOS 15, because otherwise it is easy to accidentally
refresh the page.

Since the `overscroll-behavior` CSS property wasn't supported in
Safari until version 16, this change doesn't fix the behavior in
iOS/iPadOS 15. I'm not currently sure of the best way to fix this
in a way that works in that version.