Muted tracks can have their volume modified
Previously, when a track was muted its volume slider was also
disabled. I think this was probably originally done as an
affordance to show that the track is muted, but it seems like an
unnecessary restriction.

Since no other sliders in the app are ever disabled, the "disabled
slider" CSS styles are also being removed.