Removing CSS hover styles which were never shown
The `.button-enabled:hover` styles were never shown, due to a
conflict with the `.button-hollow:enabled:hover` styles. Every
element that has the `.button-enabled` class currently also has
the `.button-hollow` class, and the `.button-hollow` hover styles
won the conflict (I believe) due to the greater specificity
provided by the `:enabled` psuedo-class.

(Confusingly, `.button-enabled` doesn't refer to whether a button
is enabled in the HTML/CSS sense, but whether it is toggled or
not. Currently all buttons with the `.button-enabled` class are
enabled in the HTML/CSS sense, but in theory they might not be).

To prevent confusion, or unintended behavior in the future, this
commit removes the `.button-enabled:hover` style declaration.

As a side benefit, after this commit there should no longer be any
`:hover` psuedo-class styles that aren't fenced inside a `@media
(hover)` media query.