Renaming `.button-enabled` -> `.button-toggled`
`.button-enabled` was a confusing CSS class name because this
class doesn't have anything to do with whether a button is enabled
in the HTML/CSS sense. Rather, it is used to style the "on" state
for a button that can be toggled on or off. (Similar to how a
checkbox can be checked or unchecked while continuously being
enabled).