Fixing border of "disabled due to error" buttons
ac2bc4914108deefb083ba71b4c9760c81784109 caused the border of
"disabled due to error" buttons to appear incorrect. While the top
and left edges had the correct light red color, the bottom and
right edges had a darker shade of red. (At least in Safari,
Firefox, and Chrome on my macOS computer; the appearance might
have been different in other browsers or OSes).

The issue was that that commit removed "solid" as the default
border style for buttons, which resulted in this type of button
not having a specific border style defined. As a result it had the
browser's default button border style, which in Safari, Firefox,
and Chrome on my computer is `outset`.