Using `webpack-dev-server` for local dev server
This allows the local server used for development to be installed
via `yarn install`, just like the app's other development
dependencies.

The new intended development model is that while making code
changes the new `yarn serve` script will be run to allow testing
the changes in a browser. When ready to deploy the changes to a
remote server, the new `yarn build` script will be run to create a
stand-alone build in the `dist/` folder.

Both of the new scripts run using Webpack's production mode.
Although this causes builds to take longer than when using
development mode, it ensures that the code being tested locally is
the same as the code that will actually be deployed. As a result
there are no longer separate `build-dev` and `build-prod` scripts.
If for some reason it's desired to use development mode anyway, it
can still be used by appending `--mode development` to either
`yarn serve` or `yarn build`.