Using `/* */` instead of `//` in CSS code comment
All of the other existing comments in the CSS code are in `/* */`
form, so this change makes this comment more consistent. Also,
although `//` comments are valid when using Sass, they are not
valid in plain CSS. Therefore this change also removes a blocker
for moving away from Sass in the future.