Upgrading dependencies
Jest v30.0.0 drops support for the `toThrowError()` function, so
its usage in tests is being replaced with the equivalent
`toThrow()`.

Sass v1.80.0 deprecates the Sass `@import` at-rule (which is
different from the CSS `@import` at-rule). Without code changes,
upgrading to this version or higher causes deprecation warnings to
appear in the build process output. If I understand things
correctly, the required code changes involve either adding a
namespace reference to each place a Sass variable is used (i.e.
`some_namespace.$var` instead of `$var`), or extracting the
variable definitions out of `main.scss` to a dedicated file. Since
the plan is to stop using Sass altogether in the future, to reduce
code churn Sass is only being upgraded to v1.79.6, which is the
most recent version before v1.80.0.